docs(terraform): OpenAPI callback for frontend client
When using the OpenAPI (rapidoc) web interface, for obtaining token
it uses /openapi/oauth-receiver.html as target for callback redirect.
( as can be seen also in common/infrastructure/src/app/http.rs#L607 )

While Cognito requires each callback page to be specified by full URL,
having just domain or top-level url (as in console-url) is not enough.

So this adds the openapi endpoint directly to allowed CallbackURLs
for the fronted client.